pyproject.toml
[project]
name = "alembic-postgress"
version = "0.1.0"
description = ""
authors = [
    {name = "Muhammad15246",email = "faizybangash05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo (>=4.13.2,<5.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.3"

